+++ Changelog.md
@@ -2,7 +2,7 @@
 
 ## master (unreleased)
 
-- Added Spain regions: Andalusia (#531).
+- Added Spain regions: Andalusia, Castile and León (#531).
 - Refactored Spain test modules (#531).
 - Fix Catalonia calendar by removing *Sant Juan* day, which does not appear to be an official holiday (#531).